pyproject.toml
[project]
name = "manifoldpy"
version = "1.1.0"
authors = [
  { name="Vincent Luczkow", email="vincent.luczkow@gmail.com" },
]
description = "Python tools for working with Manifold Markets"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "attrs==22.1.0",
    "matplotlib==3.5.2",
    "numpy==1.23.1",
    "pandas==1.5.1",
    "requests==2.28.1",
    "scipy==1.8.1"
]

[project.optional-dependencies]
dev = [
    "coverage==6.5.0",
    "mypy==0.991",
    "pytest==7.1.2",
    "pandas-stubs==1.5.1.221024",
    "Sphinx==5.3.0",
    "sphinx_rtd_theme==1.1.1",
    "types-pytz==2022.6.0.1",
    "types-requests==2.28.11.4",
    "types-urllib3==1.26.25.3"
]

[project.urls]
"Homepage" = "https://github.com/vluzko/manifold-markets-python"
"Bug Tracker" = "https://github.com/vluzko/manifold-markets-python/issues"


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
